+++ pages/index.module.css
@@ -89,16 +89,10 @@
   background-color: #fafafa;
 }
 
-.save{
-  padding: 12px 0;
-  color: #fff;
-  background-color: #10a37f;
-  border: none;
-  border-radius: 4px;
-  text-align: center;
-  cursor: pointer;
-
+.submit {
+  width: 100%;
 }
+
 
 .main .entries{
   color:#10a37f;
