+++ pages/posts/entries.js
@@ -3,6 +3,9 @@ import firebase from "../../firebase";
 import { useEffect, useState } from "react";
 import styles from "../index.module.css";
 import Link from "next/link";
+import { Typography } from 'antd';
+
+const { Title, Text } = Typography;
 
 export default function Entries() {
   const [entries, setEntries] = useState([]);
@@ -25,13 +28,14 @@ export default function Entries() {
         <title>Entries</title>
       </Head>
       <main className={styles.main}>
-        <h3>Entries</h3>
+        <Title level={3}>Entries</Title>
+        <Link href="/"> Back To Home </Link>
         {entries.map((entry) => (
           <div key={entry.time}>
-            <h2>Date: {entry.date}</h2>
-            <p>
+            <Title level={2}>Date: {entry.date}</Title>
+            <Text>
               <b>Entry:</b> {entry.entry}
-            </p>
+            </Text>
           </div>
         ))}
